refactor of codebase
