+++ playwright/.auth/state.json
@@ -2,17 +2,17 @@
   "cookies": [
     {
       "name": "oauth_state",
-      "value": "a1a50552-46b1-41f6-b271-acfc8397066c",
+      "value": "c312ac53-be36-4522-bca6-68d3f09d1ba7",
       "domain": "localhost",
       "path": "/",
-      "expires": 1770543541.786967,
+      "expires": 1770958416.169579,
       "httpOnly": true,
       "secure": false,
       "sameSite": "Lax"
     },
     {
       "name": "_cfuvid",
-      "value": "r8D4IEg9q_HkyqTaDZR8g_JssL.3h7ui0K0T1Y2AG3c-1770542942125-0.0.1.1-604800000",
+      "value": "6yK0SJD0OwbKArKmH4fxyiZ7N1ETLFfhwbvsZD6O_aE-1770957816645-0.0.1.1-604800000",
       "domain": ".workos.com",
       "path": "/",
       "expires": -1,
@@ -22,27 +22,27 @@
     },
     {
       "name": "__wuid",
-      "value": "Fe26.2*1*fb919c7051093b6a77602480e944064e4f265f59f48a0253165a667ec4bbd59f*qTIv2aaB_EFfdomq9Gw7aA*FZsIvFYJRWJJZgbKL0rIcQ5UH3HHMWhhJAp6vYh7RxO5vf1SzLF8wHmLBcUPB-idRJhqT-hoDI9RjJP5PldpnUWKgtoA9BEGh26lvbLMnAQ*1775726945104*cc683beeea1ea521da72227c6b321e5cdd6ce6eed43cb03dd1240a47373c47de*aTHGiXPv1CDm72RMLsQWfmZKYsHmbm6HXYgeulliJJ8~2",
+      "value": "Fe26.2*1*4b2ce470d094cc3c0fec098cae10407ff12210ec6460ed602f4ab44aebd2be76*iX6YUFxQIolBaZdRJvNPWg*uhquioUM1f2P3jPx47Kf2cDx6uc2S3DKNAVZDGRM_A3pyz2K8ZV4byxbzeqRrFcIAPmdvZmY5c27gwYAXfEVK-XyMMT1z32YzdiiySLeo3Q*1776141819747*7b446972e8f5b4102aa0e0b839da2c387814289c1df7b0b3bc4ac548e1c393bc*6M4Wk8JxuiCchsscwIM3__1jKFUUWX8BKo27mw2JvrU~2",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
-      "expires": 1775726945,
+      "expires": 1776141819,
       "httpOnly": false,
       "secure": true,
       "sameSite": "Lax"
     },
     {
       "name": "_dd_s",
-      "value": "aid=2411dc3e-5d59-4439-a34d-2eef18fbe0b3&rum=2&id=004f1d8f-364a-43da-81c0-8a356a340f8d&created=1770542944050&expire=1770543846812",
+      "value": "aid=5fb1943f-8d8e-4481-bef3-4bbba2f19147&rum=2&id=1f1417c2-d5f3-4246-8106-f8aca61c77a7&created=1770957818732&expire=1770958721487",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
-      "expires": 1802078946,
+      "expires": 1802493821,
       "httpOnly": false,
       "secure": false,
       "sameSite": "Strict"
     },
     {
       "name": "session",
-      "value": "aZGXBJCii21LL6A3Lpvl2Qkvb",
+      "value": "Eeb2BEsbTCM6exQY3aGcm4y7I",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
       "expires": -1,
@@ -52,40 +52,40 @@
     },
     {
       "name": "access-token",
-      "value": "eyJhbGciOiJSUzI1NiIsImtpZCI6InNzb19vaWRjX2tleV9wYWlyXzAxSzJINURFTUZUVlBEWlNLSFpWQ0RHWkM4In0.eyJhdWQiOiJodHRwczovL2FwaS53b3Jrb3MuY29tL3gvYXV0aGtpdCIsImlzcyI6Imh0dHBzOi8vYXBpLndvcmtvcy5jb20veC9hdXRoa2l0Iiwic3ViIjoidXNlcl8wMUs5Q0ZROTNZQ0EwRDhBUDg1QVNEUVdLUiIsInNpZCI6InNlc3Npb25fMDFLR1k5QVE0OTNKVjZDNDVRSDZOSkVTU0giLCJqdGkiOiIwMUtHWTlBUTVTVkM1VFY1QVJZQkoyMlYxMSIsImV4cCI6MTc3MDU0MzI0NywiaWF0IjoxNzcwNTQyOTQ3fQ.M2y-3ZY0roB9_z63h6cxTIChCRazeJ-DiukNYjzwElFdhAXGXaXPpHevjxGiRVfCrNuKiWe9sQd679QA62rzApfjhkHzQzhTi26gc6lthfl4b2cJQW4CzswKvtonLwMwydhEzc3ej6wPkE6fTUpwvHaVC-grlL7mZC8q05pOwKZvNDTCHlJI3LXP-DnMpVlbF4npb2S0ojTO4CXbdA9EjCOvO7geY-18mrb8Udz8ylvM997dj4MUp_Beu5rgPA6bo5ryG0o1JK0wu2OlujhKv1ikj6xlu8_J8WYGEMNNU6-mSJXBav93fczMXQYBIRRGmM5u-adpqLzf84yQ-w1okQ",
+      "value": "eyJhbGciOiJSUzI1NiIsImtpZCI6InNzb19vaWRjX2tleV9wYWlyXzAxSzJINURFTUZUVlBEWlNLSFpWQ0RHWkM4In0.eyJhdWQiOiJodHRwczovL2FwaS53b3Jrb3MuY29tL3gvYXV0aGtpdCIsImlzcyI6Imh0dHBzOi8vYXBpLndvcmtvcy5jb20veC9hdXRoa2l0Iiwic3ViIjoidXNlcl8wMUs5Q0ZROTNZQ0EwRDhBUDg1QVNEUVdLUiIsInNpZCI6InNlc3Npb25fMDFLSEFNWlA3WDA2V043WUNBU1A4RUdWNkYiLCJqdGkiOiIwMUtIQU1aUDk4R1BWSllTTUZKVDUwRFdYQSIsImV4cCI6MTc3MDk1ODEyMiwiaWF0IjoxNzcwOTU3ODIyfQ.1QARTdJmZNKfSXC_5fEww5-6fjXsIUoc2dKrxjBq4elcXtUUCxgLAEkeB7MlbkxKZhxim3rgZa75GGQyY3xmQFzRYgUif9wHcT85mAjkJl9ys8e4XqRaDV1WxxJFjlzn8bmzmnp-N0W4383wzEMX1MYQzy9TFTH4iwoZig5VSEruVGIeZKR_xqB4QO-R5rbPmRLaH09_rKr1-TJIWyT5WeVWTXlQjI0vvoAgMaad8Pin6bcHcOfEoE2KkqxJv4_kPP_T_a6jHDRib5oizSgqlPPyhcj4_PtjgkTVxo5Q407c0iocW1sHW-9a46aWBkus5oBmshGvjcT0LyVfiMHaxQ",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
-      "expires": 1770543247.773501,
+      "expires": 1770958122.493965,
       "httpOnly": true,
       "secure": true,
       "sameSite": "Lax"
     },
     {
       "name": "__kduid",
-      "value": "Fe26.2*1*f00ea9e85d5124a195f2633a95fc972bd0fb1fe4b87eb2447809abcc66b2094b*TZkTV7m4yJMW4FfZ1wEIgQ*Hbmgfk9TsYy6oVayb9rwG_lkUYHCOjRpNQc5PBTZgUmQhb2embZxIqB4ztByak1kbX_hR2iu8XCUS__Q6Von9HuyiKAtQYZZdxKw_fGcDmtr5ewVm6KwVNIziEsbvcbU*1802078947597*dd87ba499e87a170e2fabaeaa1864d1744871f104cf29f6abfc8569a74905945*yO4mnspiM_K-XbUQ6qx84OpXhplpyrBqinGu1Z9Ke2M~2",
+      "value": "Fe26.2*1*58feae0778f50c00f6e806ac855bb10ca8219df5e6bc26ef4a5e7591ac39941f*uNMqLPIMs6Dtpm1_Wr8b2g*b3xZKCJ8SZfdCWvZbac2v8ptWjiScK5Q5sE6LdYXRLQcTLlgfbGKzQXtsJh1lt5T8YDWVM-hja7Tr4pphQFz5_V9LtnXnQayKORmRflH1yWFEIPQBkIGAkp35zxaJyU6*1802493822303*47f18b1c7824068693af3421ed690d1d4c5e01baeb43af47ad5325b73342e232*Mp3Vi4kCXSvEvgR9clLi6FWhaWeHBwvxRs8NTKkr2Hc~2",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
-      "expires": 1802078947.773549,
+      "expires": 1802493822.493987,
       "httpOnly": true,
       "secure": true,
       "sameSite": "Lax"
     },
     {
       "name": "__recent_auth",
-      "value": "Fe26.2*1*fe102f086bfdf907fcfddff361d7b8d15ffeb2d659575cb165a1a7177b5c07d0*5jD2kA-y34HfnvSEZGcmdQ*ZtpaKGjhP2F-M-GPq78zwrIlG902fBGoTfAsqkNvWDVgmKrMo2OXHEnJLm2BLpZ1TYMe9fyF26EEvzavAeyZTw*1771147747599*398314cee2234e7ba7c87b8d16645debc77244998dfd5b9933d01e6ab15abb7c*XSH4nHgWIHCMfVNL0OIgTu6rQS3sUuiSczupz2fAspA~2",
+      "value": "Fe26.2*1*21824399c7f72b5fcff03ded18aefd1e07337507db9f8930ea597113f03c0c81*E4lWJ1PjCATtW4ZGb8JPgQ*g5jK5-07bkrkPFail74Yw4ucQrzqZ4EguSgd9wvNoyOsvcOETsVr3mXCXMJU-7OSCyIKvwzGoYNBnxBcUCKjRw*1771562622305*30e6cdb719ef0f35e22be8774c072633099dc78470f4d7ebe414860f679024e5*wRY3VayFdDc7PNZf3RcO9P8CPjFiGVQ7XKU6zGPjAnE~2",
       "domain": "fastidious-fuchsia-08-staging.authkit.app",
       "path": "/",
-      "expires": 1771147747.773581,
+      "expires": 1771562622.494,
       "httpOnly": true,
       "secure": true,
       "sameSite": "Lax"
     },
     {
       "name": "session",
-      "value": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyXzAxSzlDRlE5M1lDQTBEOEFQODVBU0RRV0tSIiwiZW1haWwiOiJzdGV2aW8ud29uZGVyQGdtYWlsLmNvbSIsIndvcmtvc0lkIjoic2Vzc2lvbl8wMUtHWTlBUTQ5M0pWNkM0NVFINk5KRVNTSCIsImlhdCI6MTc3MDU0Mjk0OCwiZXhwIjoxNzcwODg4NTQ4LCJpc3MiOiJmaXQtd29ya291dC1hcHAiLCJhdWQiOiJmaXQtd29ya291dC1hcHAifQ.obfO5IWx5qZyJreg2HU5m_FX9i3U8JZQnixUVZnKqME",
+      "value": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyXzAxSzlDRlE5M1lDQTBEOEFQODVBU0RRV0tSIiwiZW1haWwiOiJzdGV2aW8ud29uZGVyQGdtYWlsLmNvbSIsIndvcmtvc0lkIjoic2Vzc2lvbl8wMUtIQU1aUDdYMDZXTjdZQ0FTUDhFR1Y2RiIsImlhdCI6MTc3MDk1NzgyMiwiZXhwIjoxNzcxMzAzNDIyLCJpc3MiOiJmaXQtd29ya291dC1hcHAiLCJhdWQiOiJmaXQtd29ya291dC1hcHAifQ.sW5utRXDxMxXqenaAposQbvM8iqiUdHv8BOuLS7DWQs",
       "domain": "localhost",
       "path": "/",
-      "expires": 1770888548.355783,
+      "expires": 1771303422.910456,
       "httpOnly": true,
       "secure": true,
       "sameSite": "Lax"
@@ -101,7 +101,7 @@
         },
         {
           "name": "auth_user",
-          "value": "{\"id\":\"user_01K9CFQ93YCA0D8AP85ASDQWKR\",\"email\":\"stevio.wonder@gmail.com\",\"name\":\"stevio.wonder\",\"cachedAt\":1770542948629}"
+          "value": "{\"id\":\"user_01K9CFQ93YCA0D8AP85ASDQWKR\",\"email\":\"stevio.wonder@gmail.com\",\"name\":\"stevio.wonder\",\"cachedAt\":1770957823107}"
         }
       ]
     }
