+++ src/components/ui/Card.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 
 import { cn } from '~/lib/cn'
 
-function Card({ className, variant = 'surface', ...props }: React.ComponentProps<'div'> & { variant?: 'surface' | 'elevated' | 'tinted' }) {
+function Card({ className, variant = 'surface', overflow = 'hidden', ...props }: React.ComponentProps<'div'> & { variant?: 'surface' | 'elevated' | 'tinted'; overflow?: 'visible' | 'hidden' | 'auto' | 'clip' | 'scroll' }) {
   const variantStyles = {
     surface: 'border border-border/60 bg-card shadow-xs',
     elevated: 'border border-border/40 bg-card shadow-md',
@@ -13,10 +13,11 @@ function Card({ className, variant = 'surface', ...props }: React.ComponentProps
     <div
       data-slot="card"
       className={cn(
-        'flex flex-col gap-6 rounded-xl py-5 transition-shadow duration-200 overflow-hidden',
+        'flex flex-col gap-6 rounded-xl py-5 transition-shadow duration-200',
         variantStyles[variant],
         className,
       )}
+      style={{ overflow }}
       {...props}
     />
   )
