+++ src/lib/db/exercise/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './repository';
