+++ src/lib/db/workout/index.ts
@@ -1,2 +1,7 @@
 export * from './types';
-export * from './repository';
+export * from './workouts';
+export * from './exercises';
+export * from './sets';
+export * from './history';
+export * from './stats';
+export { calculateE1RM } from '~/lib/domain/stats/calculations';
