+++ tests/unit/exercise.spec.ts
@@ -56,6 +56,12 @@ describe('Exercise CRUD Operations', () => {
     createDbMock = vi.fn(() => mockDrizzleDb);
     vi.doMock('../../src/lib/db/index', () => ({
       createDb: createDbMock,
+      getDb: vi.fn((dbOrTx) => {
+        if (dbOrTx && typeof dbOrTx === 'object' && 'transaction' in dbOrTx) {
+          return dbOrTx;
+        }
+        return mockDrizzleDb;
+      }),
     }));
   });
